Moves modules/drafts => modules/draftMessage - remove modules/drafts
