+++ src/frontend/web_application/src/modules/draftMessage/components/DiscussionDraft/index.jsx
