+++ src/frontend/web_application/src/modules/draftMessage/components/DiscussionTextarea/presenter.jsx
