+++ src/frontend/web_application/src/modules/draftMessage/components/DiscussionTextarea/style.scss
