+++ src/frontend/web_application/src/modules/draftMessage/components/DraftMessageActionsContainer/style.scss
