+++ src/frontend/web_application/src/modules/draftMessage/components/NewDraftForm/style.scss
