+++ src/frontend/web_application/src/modules/draftMessage/components/Recipient/index.js
