+++ src/frontend/web_application/src/modules/draftMessage/components/Recipient/presenter.spec.jsx
