+++ src/frontend/web_application/src/modules/draftMessage/components/RecipientList/style.scss
