+++ src/frontend/web_application/src/modules/draftMessage/components/ReplyForm/presenter.spec.jsx
