+++ src/frontend/web_application/src/modules/draftMessage/components/ReplyForm/style.scss
