+++ src/frontend/web_application/src/modules/draftMessage/index.js
@@ -1,2 +1,5 @@
 export * from './actions/saveDraft';
 export * from './actions/sendDraft';
+export { default as NewDraftForm } from './components/NewDraftForm';
+export { default as DraftMessageActionsContainer } from './components/DraftMessageActionsContainer';
+export { default as ReplyForm } from './components/ReplyForm';
