--- src/frontend/web_application/src/modules/drafts/index.jsx
@@ -1,3 +1,0 @@
-export { default as NewDraftForm } from './components/NewDraftForm';
-export { default as DraftMessageActionsContainer } from './components/DraftMessageActionsContainer';
-export { default as ReplyForm } from './components/ReplyForm';
